% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{add_lookup_users_info_to_nodes_in_graph}
\alias{add_lookup_users_info_to_nodes_in_graph}
\title{Fetches the user's lookup_users data then updates their info in the graph.}
\usage{
add_lookup_users_info_to_nodes_in_graph(user_ids, cache)
}
\arguments{
\item{user_ids}{a vector of user_ids. do NOT pass screen names}

\item{cache}{the cache to interface with}
}
\value{
tibble of user data
}
\description{
Fetches the user's lookup_users data then updates their info in the graph.
}
