% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{clear____db}
\alias{clear____db}
\title{HELPER FUNCTION USED TO CLEAR THE DATABASE FOR TESTING,
REMOVE THIS FUNCTION ONCE TESTING IS DONE}
\usage{
clear____db()
}
\description{
HELPER FUNCTION USED TO CLEAR THE DATABASE FOR TESTING,
REMOVE THIS FUNCTION ONCE TESTING IS DONE
}
