% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friends.R
\name{db_connect_friends}
\alias{db_connect_friends}
\title{Gets the friends for the given user_id and creates the edges in the graph.
This function is only ever called by add_new_friends.}
\usage{
db_connect_friends(user_ids, n)
}
\arguments{
\item{user_ids}{user_ids who are already in the database and
do not have friend edge data}
}
\value{
a nx2 tibble where the <from> column is user_id and the <to> column
is the user_id of user_id's friends
}
\description{
Gets the friends for the given user_id and creates the edges in the graph.
This function is only ever called by add_new_friends.
}
