% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{docker_bulk_connect_nodes}
\alias{docker_bulk_connect_nodes}
\title{This function creates edges en masse between all the nodes provided in the
tbl argument.}
\usage{
docker_bulk_connect_nodes(tbl, cache)
}
\arguments{
\item{tbl}{tibble containing columns for 'to' and 'from' consisting of user_ids}
}
\value{
the same tibble edge list that was provided as an argument
}
\description{
This function creates edges en masse between all the nodes provided in the
tbl argument.
}
