% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j.R
\name{docker_bulk_merge_users}
\alias{docker_bulk_merge_users}
\title{Merges a batch of nodes to the graph with nothing but user_id's}
\usage{
docker_bulk_merge_users(user_ids, cache)
}
\arguments{
\item{user_ids}{a vector of user_ids to generate MERGE queries for}

\item{cache}{the cache to interface with}
}
\description{
Merges a batch of nodes to the graph with nothing but user_id's
}
