% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{empty_edge_list}
\alias{empty_edge_list}
\title{Create an empty table of following information}
\usage{
empty_edge_list()
}
\value{
an empty 2-column tibble used as a placeholder for
when user data is not available
}
\description{
Create an empty table of following information
}
\keyword{internal}
