% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{empty_lookup}
\alias{empty_lookup}
\title{An empty version of the tibble returned by \code{rtweet}'s lookup_users
fuctions. Used when binding with empty query results to avoid errors.}
\usage{
empty_lookup()
}
\value{
an empty tibble with columns named after all of the lookup_user
properties.  Used when user data is unavailable
}
\description{
An empty version of the tibble returned by \code{rtweet}'s lookup_users
fuctions. Used when binding with empty query results to avoid errors.
}
\keyword{internal}
