% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{empty_user_edges}
\alias{empty_user_edges}
\title{An empty edge list, used when binding with empty query results to
avoid errors.}
\usage{
empty_user_edges()
}
\value{
an empty 2-column tibble used as a placeholder for
when user data is not available
}
\description{
An empty edge list, used when binding with empty query results to
avoid errors.
}
\keyword{internal}
