% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followers.R
\name{fetch_followers}
\alias{fetch_followers}
\title{Fetches the followers of the provided user ID's via \code{rtweet}.}
\usage{
fetch_followers(user_ids, n)
}
\value{
an nx2 tibble edge list
}
\description{
Fetches the followers of the provided user ID's via \code{rtweet}.
}
