% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friends.R
\name{fetch_friends}
\alias{fetch_friends}
\title{Uses rtweet to fetch the friend list of the users provided in user_ids}
\usage{
fetch_friends(user_ids, n)
}
\arguments{
\item{user_ids}{a vector of user ids}
}
\value{
an nx2 tibble edge list
}
\description{
Uses rtweet to fetch the friend list of the users provided in user_ids
}
