% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{fetch_lookup}
\alias{fetch_lookup}
\title{Takes in a vector of user_ids and returns their lookup_users info.}
\usage{
fetch_lookup(user_ids)
}
\arguments{
\item{user_ids}{vector of user_ids}
}
\value{
tibble of user data
}
\description{
Takes in a vector of user_ids and returns their lookup_users info.
}
