% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{fetch_lookup_update}
\alias{fetch_lookup_update}
\title{Fetches the user's lookup_users data then updates their info in the graph.}
\usage{
fetch_lookup_update(user_ids, cache)
}
\arguments{
\item{user_ids}{a vector of user_ids}

\item{cache}{the cache to interface with}
}
\value{
tibble of user data
}
\description{
Fetches the user's lookup_users data then updates their info in the graph.
}
