% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followers.R
\name{follower_sampling_status}
\alias{follower_sampling_status}
\title{Checks whether follower data has already been sampled for the provided
vector of users.}
\usage{
follower_sampling_status(user_ids)
}
\arguments{
\item{user_ids}{to fetch the sampling status for}
}
\value{
a list of all users who either (1) are not currently in the
graph, (2) are in the graph but their followers have not been sampled,
(3) are in the graph and have sampled followers
}
\description{
Checks whether follower data has already been sampled for the provided
vector of users.
}
