% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache}
\alias{get_cache}
\title{Retrieves a previously generated cache object.}
\usage{
get_cache(cache_name)
}
\arguments{
\item{cache_name}{the name of the cache to use}
}
\value{
a list object containing information pertaining to the associated
Docker container and Neo4j instance
}
\description{
Retrieves a previously generated cache object.
}
\keyword{internal}
