% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followers.R
\name{get_followers}
\alias{get_followers}
\title{Fetches the follows of all of the users in a provided vector. Returns
the follower relationships in the form of an edge list.}
\usage{
get_followers(user_ids, cache, n = 150)
}
\arguments{
\item{user_ids}{A character vector of user ids (never screen names)}
}
\value{
A tibble where each row corresponds to a follower relationship
from the user in the 'from' column to the user in to 'to' column
}
\description{
Fetches the follows of all of the users in a provided vector. Returns
the follower relationships in the form of an edge list.
}
