% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friends.R
\name{get_friends}
\alias{get_friends}
\title{Fetches the friends of each of the users contained in the user_ids vector.
Returns the friend relationships via a tibble edge list.}
\usage{
get_friends(user_ids, n = 150)
}
\arguments{
\item{user_ids}{A character vector of user ids (never screen names)}
}
\value{
A tibble where each row corresponds to a follower relationship
from the user in the 'from' column to the user in to 'to' column
}
\description{
Fetches the friends of each of the users contained in the user_ids vector.
Returns the friend relationships via a tibble edge list.
}
