% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{load_cache}
\alias{load_cache}
\title{Starts the Docker container and Neo4j instance for the given cache}
\usage{
load_cache(cache)
}
\arguments{
\item{cache}{the cache object that the user wishes to start}
}
\description{
Starts the Docker container and Neo4j instance for the given cache
}
