% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{lookup_users}
\alias{lookup_users}
\title{Looks up information pertaining to a vector of provided users.  Fetches cached
data for users who have already been looked up, and fetches new data for users
who have not bee looked up yet.}
\usage{
lookup_users(user_ids, cache)
}
\arguments{
\item{users}{A character vector of user ids (never screen names)}
}
\value{
A tibble where each row corresponds to a User and each column
to one of the User properties. If a user cannot be sampled, should
return nothing for that user. If no users can be sampled, should
return an empty tibble with appropriate columns.
}
\description{
Looks up information pertaining to a vector of provided users.  Fetches cached
data for users who have already been looked up, and fetches new data for users
who have not bee looked up yet.
}
