% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{merge_fetch_lookup_update}
\alias{merge_fetch_lookup_update}
\title{Creates nodes for the given user_ids in the graph, then fetches their
info and updates that info in the graph.}
\usage{
merge_fetch_lookup_update(user_ids, cache)
}
\arguments{
\item{user_ids}{a vector of user_ids}

\item{cache}{the cache to interface with}
}
\value{
tibble of user data
}
\description{
Creates nodes for the given user_ids in the graph, then fetches their
info and updates that info in the graph.
}
