% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followers.R
\name{merge_then_fetch_connect_followers}
\alias{merge_then_fetch_connect_followers}
\title{user_ids is a list of COMPLETELY NEW users. This function performs the following:
\enumerate{
\item Fetch the followers of each user (call these main users) listed in user_ids (call these blank followers)
\item MERGE nodes for main users and blank followers (each of these nodes will only contain a user_id field)
\item Create edges between main users and their respective blank followers
\item Set the sampled_followers_at property for nodes that were sampled
}}
\usage{
merge_then_fetch_connect_followers(user_ids, n, cache)
}
\arguments{
\item{user_ids}{a list of user_ids to add follower edges to the db for}

\item{n}{how many followers to sample at a time for each user}
}
\value{
a 2-column tibble edge list from user_ids to their followers
}
\description{
user_ids is a list of COMPLETELY NEW users. This function performs the following:
\enumerate{
\item Fetch the followers of each user (call these main users) listed in user_ids (call these blank followers)
\item MERGE nodes for main users and blank followers (each of these nodes will only contain a user_id field)
\item Create edges between main users and their respective blank followers
\item Set the sampled_followers_at property for nodes that were sampled
}
}
