% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friends.R
\name{merge_then_fetch_connect_friends}
\alias{merge_then_fetch_connect_friends}
\title{user_ids is a list of COMPLETELY NEW users. This function performs the following:
\enumerate{
\item Fetch the friends of each user (call these main users) listed in user_ids (call these blank friends)
\item MERGE nodes for main users and blank friends (each of these nodes will only contain a user_id field)
\item Create edges between main users and their respective blank friends
\item Set the sampled_friends_at property for nodes that were sampled
}}
\usage{
merge_then_fetch_connect_friends(user_ids, n)
}
\arguments{
\item{user_ids}{a list of user_ids to add friend edges to the db for}

\item{n}{how many friends to sample at a time for each user}
}
\value{
a 2-column tibble edge list from user_ids to their friends
}
\description{
user_ids is a list of COMPLETELY NEW users. This function performs the following:
\enumerate{
\item Fetch the friends of each user (call these main users) listed in user_ids (call these blank friends)
\item MERGE nodes for main users and blank friends (each of these nodes will only contain a user_id field)
\item Create edges between main users and their respective blank friends
\item Set the sampled_friends_at property for nodes that were sampled
}
}
