% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{merge_users}
\alias{merge_users}
\title{TODO: Rename this function. This function is used both to update present users
and add new users to the db.}
\usage{
merge_users(user_ids, lookup, n = 150)
}
\arguments{
\item{user_ids}{the user_ids to update}

\item{lookup}{should new Twitter profile data be updated for each user_id?}

\item{n}{how many friends/followers should be looked up at a time if
the respective argument is set to TRUE?}
}
\value{
The tibble of user data, with one row for each (accessible)
user in \code{users} and one column for each property of \code{User} nodes
in the graph database.
}
\description{
TODO: Rename this function. This function is used both to update present users
and add new users to the db.
}
