% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nc_activate_cache}
\alias{nc_activate_cache}
\title{Starts the Docker container and Neo4j instance for the given cache}
\usage{
nc_activate_cache(cache_name)
}
\arguments{
\item{cache_name}{the name of the cache and associated Docker container}
}
\description{
Starts the Docker container and Neo4j instance for the given cache
}
\keyword{internal}
