% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nc_browse}
\alias{nc_browse}
\title{Interact with a cached graph via the Neo4J browser interface}
\usage{
nc_browse(cache_name)
}
\arguments{
\item{cache_name}{the name of the cache and associated Docker container}
}
\description{
Interact with a cached graph via the Neo4J browser interface
}
