% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nc_cache_exists}
\alias{nc_cache_exists}
\title{Check if a cache with a given name exists}
\usage{
nc_cache_exists(cache_name)
}
\arguments{
\item{cache_name}{}
}
\value{

}
\description{
Check if a cache with a given name exists
}
