% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nc_create_cache}
\alias{nc_create_cache}
\title{Create a cache object}
\usage{
nc_create_cache(
  cache_name,
  neo4j_pass = "password",
  http_port = 7474,
  bolt_port = 7687,
  url = NULL
)
}
\arguments{
\item{cache_name}{the name of the cache and associated Docker container}

\item{neo4j_pass}{the password for the Neo4j instance}

\item{http_port}{the port that Neo4j will use from inside of the Docker container for its HTTP connection}

\item{bolt_port}{the port that Neo4j will use from inside of the Docker container for its Bolt connection}

\item{url}{the url used to connect to the Neo4j database}
}
\description{
Creates and saves a new cache instance with the given information.  Also
generates a new Docker container with the given information if one does
not already exist.
}
\examples{

example_cache <- nc_create_cache("example")
example_cache

nc_sitrep()

nc_destroy_cache(example_cache)

nc_sitrep()

}
