% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nc_destroy_cache}
\alias{nc_destroy_cache}
\title{Removes the Docker container and cache save file corresponding to the given
cache}
\usage{
nc_destroy_cache(cache_name)
}
\arguments{
\item{cache}{the cache object that the user wishes to start}
}
\description{
Removes the Docker container and cache save file corresponding to the given
cache
}
