% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nc_empty_cache}
\alias{nc_empty_cache}
\title{Remove all data from a cache}
\usage{
nc_empty_cache(cache_name, check_with_me_first = TRUE)
}
\arguments{
\item{cache}{}
}
\value{

}
\description{
Remove all data from a cache
}
