% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j.R
\name{nc_export_all_follows}
\alias{nc_export_all_follows}
\title{Export all follows from the Neo4J database to a csv}
\usage{
nc_export_all_follows(cache_name, local_path)
}
\arguments{
\item{cache_name}{the name of the cache and associated Docker container}

\item{local_path}{path of csv}
}
\description{
Export all follows from the Neo4J database to a csv
}
