% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friends.R
\name{nc_get_friends}
\alias{nc_get_friends}
\title{Fetches the friends of each of the users contained in the user_ids vector.
Returns the friend relationships via a tibble edge list.}
\usage{
nc_get_friends(
  user_ids,
  cache_name,
  n = 5000,
  retryonratelimit = TRUE,
  cursor = "-1",
  verbose = TRUE,
  token = NULL
)
}
\arguments{
\item{n}{Desired number of results to return. Results are downloaded
in pages when \code{n} is large; the default value will download a single
page. Set \code{n = Inf} to download as many results as possible.

The Twitter API rate limits the number of requests you can perform
in each 15 minute period. The easiest way to download more than that is
to use \code{retryonratelimit = TRUE}.

You are not guaranteed to get exactly \code{n} results back. You will get
fewer results when tweets have been deleted or if you hit a rate limit.
You will get more results if you ask for a number of tweets that's not
a multiple of page size, e.g. if you request \code{n = 150} and the page
size is 200, you'll get 200 results back.}

\item{retryonratelimit}{If \code{TRUE}, and a rate limit is exhausted, will wait
until it refreshes. Most twitter rate limits refresh every 15 minutes.
If \code{FALSE}, and the rate limit is exceeded, the function will terminate
early with a warning; you'll still get back all results received up to
that point. The default value, \code{NULL}, consults the option
\code{rtweet.retryonratelimit} so that you can globally set it to \code{TRUE},
if desired.

If you expect a query to take hours or days to perform, you should not
rely soley on \code{retryonratelimit} because it does not handle other common
failure modes like temporarily losing your internet connection.}

\item{cursor}{Which page of results to return. The default will return
the first page; you can supply the result from a previous call to
continue pagination from where it left off.}

\item{verbose}{Show progress bars and other messages indicating current
progress?}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[rtweet:auth_as]{auth_as()}} for details.}

\item{cache}{The name of the cache to save data in.}
}
\value{
A tibble where each row corresponds to a follower relationship
from the user in the 'from' column to the user in to 'to' column
}
\description{
Fetches the friends of each of the users contained in the user_ids vector.
Returns the friend relationships via a tibble edge list.
}
