% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{nc_lookup_users}
\alias{nc_lookup_users}
\title{Get (cached) information about Twitter users}
\usage{
nc_lookup_users(user_ids, cache_name = NULL)
}
\arguments{
\item{user_ids}{Character vector of Twitter user ids. \strong{DO NOT}
pass Twitter screen names, as \code{user_ids} are taken to be unique
node identifiers.}

\item{cache}{A \code{cache} object. See \code{\link[=nc_create_cache]{nc_create_cache()}}.}
}
\value{
A tibble where each row corresponds to a User and each column
to one of the User properties. If a user cannot be sampled, should
return nothing for that user. If no users can be sampled, should
return an empty tibble with appropriate columns.
}
\description{
Looks up information pertaining to a vector of provided users. Retrieves
cached data for users who have already been looked up, and fetches new
data for users who have not bee looked up yet from the Twitter API.
}
