% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{remove_cache}
\alias{remove_cache}
\title{Removes the Docker container and cache save file corresponding to the given
cache}
\usage{
remove_cache(cache)
}
\arguments{
\item{cache}{the cache object that the user wishes to start}
}
\description{
Removes the Docker container and cache save file corresponding to the given
cache
}
