% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{start_neo4j}
\alias{start_neo4j}
\title{Starts the Neo4j Docker container; creates and sets up the container if it
does not already exist.}
\usage{
start_neo4j()
}
\description{
Starts the Neo4j Docker container; creates and sets up the container if it
does not already exist.
}
