% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R, R/testing.R
\name{stop_neo4j}
\alias{stop_neo4j}
\title{Stops the Neo4j Docker container.}
\usage{
stop_neo4j()

stop_neo4j()
}
\description{
Stops the Neo4j Docker container.

stop_neo4j but used exclusively in unit testing
}
