% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sup4j}
\alias{sup4j}
\title{Sends CYPHER queries to a given connexion object while suppressing output
messages that call_neo4j throws.}
\usage{
sup4j(query, con = get_connexion())
}
\arguments{
\item{query}{the CYPHER query to be passed to call_neo4j}

\item{con}{the neo4j connection object to be passed to call_neo4j}
}
\value{
the return value from call_neo4j
}
\description{
Sends CYPHER queries to a given connexion object while suppressing output
messages that call_neo4j throws.
}
\keyword{internal}
