% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{unload_cache}
\alias{unload_cache}
\title{Stops the Docker container and Neo4j instance for the given cache}
\usage{
unload_cache(cache)
}
\arguments{
\item{cache}{the cache object that the user wishes to start}
}
\description{
Stops the Docker container and Neo4j instance for the given cache
}
