% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{add_lookup_users_info_to_nodes_in_graph}
\alias{add_lookup_users_info_to_nodes_in_graph}
\title{Fetches the user's lookup_users data then updates their info in the graph.}
\usage{
add_lookup_users_info_to_nodes_in_graph(
  users,
  token,
  retryonratelimit,
  verbose,
  cache
)
}
\arguments{
\item{users}{User id or screen name of target user.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[rtweet:auth_as]{auth_as()}} for details.}

\item{retryonratelimit}{If \code{TRUE}, and a rate limit is exhausted, will wait
until it refreshes. Most Twitter rate limits refresh every 15 minutes.
If \code{FALSE}, and the rate limit is exceeded, the function will terminate
early with a warning; you'll still get back all results received up to
that point. The default value, \code{NULL}, consults the option
\code{rtweet.retryonratelimit} so that you can globally set it to \code{TRUE},
if desired.

If you expect a query to take hours or days to perform, you should not
rely soley on \code{retryonratelimit} because it does not handle other common
failure modes like temporarily losing your internet connection.}

\item{verbose}{Show progress bars and other messages indicating current
progress?}

\item{cache}{A \code{cache} object. See \code{\link[=nc_create_cache]{nc_create_cache()}}.}
}
\value{
tibble of user data
}
\description{
Fetches the user's lookup_users data then updates their info in the graph.
}
