% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j.R
\name{db_add_new_users}
\alias{db_add_new_users}
\title{Merges a batch of nodes to the graph with nothing but user_id's}
\usage{
db_add_new_users(users, cache)
}
\arguments{
\item{users}{a vector of users to generate MERGE queries for}

\item{cache}{the cache to interface with}
}
\description{
Merges a batch of nodes to the graph with nothing but user_id's
}
