% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_get_followers}
\alias{db_get_followers}
\title{Gets the followers for the given user that already exist in the DB.}
\usage{
db_get_followers(user_ids, cache)
}
\arguments{
\item{user_ids}{a list of user_ids who are already in the DB and
already have follower edge data}
}
\value{
a 2-column tibble edge list with entries from the users in user_ids
to their followers
}
\description{
Gets the followers for the given user that already exist in the DB.
}
