% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friends.R
\name{db_get_friends}
\alias{db_get_friends}
\title{Gets the friends for the given user that already exist in the DB.}
\usage{
db_get_friends(users, cache)
}
\arguments{
\item{users}{Screen name or user ID of target user from which the
user IDs of friends (accounts followed BY target user) will be
retrieved.}

\item{cache}{The cache to store information in. Must be a cache, not a
cache name.}
}
\value{
a 2-column tibble edge list with entries from the users in users
to their friends
}
\description{
Gets the friends for the given user that already exist in the DB.
}
