% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_get_friends}
\alias{db_get_friends}
\title{Gets the friends for the given user that already exist in the DB.}
\usage{
db_get_friends(user_ids, cache)
}
\arguments{
\item{user_ids}{a list of user_ids who are already in the DB and
already have friend edge data}

\item{cache}{the cache to interface with}
}
\value{
a 2-column tibble edge list with entries from the users in user_ids
to their friends
}
\description{
Gets the friends for the given user that already exist in the DB.
}
