% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{db_lookup_users}
\alias{db_lookup_users}
\title{Looks up users that are already in the database.}
\usage{
db_lookup_users(user_ids)
}
\arguments{
\item{user_ids}{list of user_ids to fetch existing lookup_user data for in
the db}
}
\value{
a tibble with any existing data for user_ids
}
\description{
Looks up users that are already in the database.
}
