% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_users.R
\name{db_lookup_users}
\alias{db_lookup_users}
\title{Looks up users that are already in the database.}
\usage{
db_lookup_users(users, cache)
}
\arguments{
\item{users}{list of users to fetch existing lookup_user data for in the db}

\item{cache}{the cache to interface with}
}
\value{
a tibble with any existing data for users
}
\description{
Looks up users that are already in the database.
}
