% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{empty_lookup}
\alias{empty_lookup}
\title{Create an empty table of user data}
\usage{
empty_lookup()
}
\value{
an empty tibble with columns named after all of the lookup_user
properties.  Used when user data is unavailable
}
\description{
Create an empty table of user data
}
\keyword{internal}
