% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followers.R
\name{fetch_followers}
\alias{fetch_followers}
\title{Fetches the followers of the provided user ID's via \code{rtweet}.}
\usage{
fetch_followers(user_ids, n)
}
\arguments{
\item{user_ids}{vector of user_ids whose followers are being fetches}

\item{n}{the number of followers to fetch for each user_id}
}
\value{
an nx2 tibble edge list
}
\description{
Fetches the followers of the provided user ID's via \code{rtweet}.
}
