% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_connexion}
\alias{get_connexion}
\title{Get the connexion object that facilitates access to the Neo4j database}
\usage{
get_connexion(cache)
}
\arguments{
\item{cache}{the cache to interface with}
}
\value{
connexion object
}
\description{
Get the connexion object that facilitates access to the Neo4j database
}
