% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nc_deactivate_cache}
\alias{nc_deactivate_cache}
\title{Stops the Docker container and Neo4j instance for the given cache}
\usage{
nc_deactivate_cache(cache_name)
}
\arguments{
\item{cache_name}{the name of the cache and associated Docker container}
}
\description{
Stops the Docker container and Neo4j instance for the given cache
}
\keyword{internal}
