% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nc_empty_cache}
\alias{nc_empty_cache}
\title{Remove all data from a cache}
\usage{
nc_empty_cache(cache_name, check_with_me_first = TRUE)
}
\arguments{
\item{cache_name}{the name of the cache and associated Docker container}

\item{check_with_me_first}{Whether or not to confirm choice to empty
cache via CLI menu before clearing the cache}
}
\description{
Remove all data from a cache
}
